move stats in bucket_stats
