+++ bucket_map/src/bucket_storage.rs
@@ -1,4 +1,5 @@
 use crate::bucket_map::MaxSearch;
+use crate::bucket_stats::BucketStats;
 use memmap2::MmapMut;
 use rand::{thread_rng, Rng};
 use solana_measure::measure::Measure;
@@ -10,24 +11,8 @@ use std::path::PathBuf;
 use std::sync::Arc;
 use std::sync::{
     atomic::{AtomicU64, Ordering},
-    Mutex,
 };
 
-#[derive(Debug, Default)]
-pub struct BucketStats {
-    pub resizes: AtomicU64,
-    pub max_size: Mutex<u64>,
-    pub resize_us: AtomicU64,
-    pub new_file_us: AtomicU64,
-    pub flush_file_us: AtomicU64,
-    pub mmap_us: AtomicU64,
-}
-
-#[derive(Debug, Default, Clone)]
-pub struct BucketMapStats {
-    pub index: Arc<BucketStats>,
-    pub data: Arc<BucketStats>,
-}
 
 /*
 1	2
